+++ MekHQ/src/mekhq/CampaignPreset.java
@@ -64,9 +64,9 @@ import megamek.common.preference.PreferenceManager;
 import megamek.common.util.sorter.NaturalOrderComparator;
 import megamek.logging.MMLogger;
 import megamek.utilities.xml.MMXMLUtility;
-import mekhq.campaign.CampaignOptions;
-import mekhq.campaign.io.CampaignOptionsMarshaller;
-import mekhq.campaign.io.CampaignOptionsUnmarshaller;
+import mekhq.campaign.campaignOptions.CampaignOptions;
+import mekhq.campaign.campaignOptions.CampaignOptionsMarshaller;
+import mekhq.campaign.campaignOptions.CampaignOptionsUnmarshaller;
 import mekhq.campaign.personnel.PersonnelOptions;
 import mekhq.campaign.personnel.SpecialAbility;
 import mekhq.campaign.personnel.ranks.RankSystem;
