+++ MekHQ/src/mekhq/campaign/CampaignSummary.java
@@ -53,6 +53,7 @@ import java.util.List;
 import megamek.common.Entity;
 import megamek.common.Infantry;
 import megamek.common.UnitType;
+import mekhq.campaign.campaignOptions.CampaignOptions;
 import mekhq.campaign.mission.Mission;
 import mekhq.campaign.mission.enums.MissionStatus;
 import mekhq.campaign.personnel.Person;
