+++ MekHQ/src/mekhq/campaign/personnel/lifeEvents/CommandersDayAnnouncement.java
@@ -45,7 +45,7 @@ import javax.swing.ImageIcon;
 import megamek.common.annotations.Nullable;
 import megamek.utilities.ImageUtilities;
 import mekhq.campaign.Campaign;
-import mekhq.campaign.CampaignOptions;
+import mekhq.campaign.campaignOptions.CampaignOptions;
 import mekhq.campaign.personnel.Person;
 import mekhq.gui.baseComponents.immersiveDialogs.ImmersiveDialogSimple;
