+++ MekHQ/src/mekhq/gui/CommandCenterTab.java
@@ -55,8 +55,8 @@ import megamek.common.event.Subscribe;
 import mekhq.MHQOptionsChangedEvent;
 import mekhq.MekHQ;
 import mekhq.campaign.Campaign;
-import mekhq.campaign.CampaignOptions;
 import mekhq.campaign.CampaignSummary;
+import mekhq.campaign.campaignOptions.CampaignOptions;
 import mekhq.campaign.event.*;
 import mekhq.campaign.finances.FinancialReport;
 import mekhq.campaign.finances.Money;
